fix qform validation add toast

+++ src/app/dashboard/add-manage/components/quiz-summary-table/quiz-summary-table.component.ts
@@ -1,11 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { QuestionsapiService } from '../../services/questionsapi.service';
 import { StateService } from '../../services/state.service';
+import { ConfirmationService, MessageService } from 'primeng/api';
 
 @Component({
   selector: 'app-quiz-summary-table',
   templateUrl: './quiz-summary-table.component.html',
   styleUrl: './quiz-summary-table.component.scss',
+  providers: [ConfirmationService, MessageService],
 })
 export class QuizSummaryTableComponent implements OnInit {
   questions: any[] = [];
@@ -13,12 +15,13 @@ export class QuizSummaryTableComponent implements OnInit {
   constructor(
     private questionsapiService: QuestionsapiService,
     private readonly _stateService: StateService,
-    private readonly _questionApiService: QuestionsapiService
+    private readonly _questionApiService: QuestionsapiService,
+    private messageService: MessageService,
+    private confirmationService: ConfirmationService
   ) {}
 
   ngOnInit(): void {
     this.loadquestiondata();
-
     this._stateService.updateMsg$.subscribe({
       next: (res: any) => {
         if (res) {
@@ -31,14 +34,10 @@ export class QuizSummaryTableComponent implements OnInit {
   }
 
   loadquestiondata() {
-    this.loading = true;
-
     this.questionsapiService.getQuestionsList().subscribe((response) => {
       this.questions = response;
       console.log('this is response:', response);
     });
-
-    this.loading = false;
   }
 
   edit(question: any) {
@@ -46,12 +45,49 @@ export class QuizSummaryTableComponent implements OnInit {
     this._stateService.sendMessage(question);
   }
 
-  delete(id: number) {
-    console.log(`question id to delete: ${id}`);
-    this._questionApiService.delete(id).subscribe({
-      next: (response: any) => {
-        console.log(response);
+  delete(event: Event, id: number) {
+    this.confirmationService.confirm({
+      target: event.target as EventTarget,
+      message: 'Do you want to delete this record?',
+      header: 'Delete Confirmation',
+      icon: 'pi pi-info-circle',
+      acceptButtonStyleClass: 'p-button-danger p-button-text',
+      rejectButtonStyleClass: 'p-button-text p-button-text',
+      acceptIcon: 'none',
+      rejectIcon: 'none',
+
+      accept: () => {
+        this._questionApiService.delete(id).subscribe({
+          next: (response: any) => {
+            console.log(response);
+            setTimeout(() => {
+              this.loadquestiondata();
+            }, 2000);
+          },
+        });
+        this.messageService.add({
+          severity: 'info',
+          summary: 'Confirmed',
+          detail: 'Record deleted',
+          life: 1000,
+        });
+      },
+      reject: () => {
+        this.messageService.add({
+          severity: 'error',
+          summary: 'Rejected',
+          detail: 'You have rejected',
+          life: 500,
+        });
       },
     });
   }
+  // delete(id: number) {
+  //   console.log(`question id to delete: ${id}`);
+  //   this._questionApiService.delete(id).subscribe({
+  //     next: (response: any) => {
+  //       console.log(response);
+  //     },
+  //   });
+  // }
 }
